# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bin Yasin Travels API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets/group:
    get:
      operationId: listGroupTickets
      tags: [tickets]
      summary: List group tickets
      parameters:
        - name: country
          in: query
          schema:
            type: string
        - name: airline
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of group tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupTicket"
    post:
      operationId: createGroupTicket
      tags: [tickets]
      summary: Create group ticket (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupTicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupTicket"
  /tickets/group/{id}:
    get:
      operationId: getGroupTicket
      tags: [tickets]
      summary: Get group ticket by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupTicket"
    patch:
      operationId: updateGroupTicket
      tags: [tickets]
      summary: Update group ticket (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupTicketUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupTicket"
    delete:
      operationId: deleteGroupTicket
      tags: [tickets]
      summary: Delete group ticket (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tickets/umrah:
    get:
      operationId: listUmrahTickets
      tags: [tickets]
      summary: List Umrah tickets
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UmrahTicket"
    post:
      operationId: createUmrahTicket
      tags: [tickets]
      summary: Create Umrah ticket (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UmrahTicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UmrahTicket"
  /tickets/umrah/{id}:
    patch:
      operationId: updateUmrahTicket
      tags: [tickets]
      summary: Update Umrah ticket (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UmrahTicketInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UmrahTicket"
    delete:
      operationId: deleteUmrahTicket
      tags: [tickets]
      summary: Delete Umrah ticket (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /packages/umrah:
    get:
      operationId: listUmrahPackages
      tags: [packages]
      summary: List Umrah packages
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UmrahPackage"
    post:
      operationId: createUmrahPackage
      tags: [packages]
      summary: Create Umrah package (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UmrahPackageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UmrahPackage"
  /packages/umrah/{id}:
    patch:
      operationId: updateUmrahPackage
      tags: [packages]
      summary: Update Umrah package (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UmrahPackageInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UmrahPackage"
    delete:
      operationId: deleteUmrahPackage
      tags: [packages]
      summary: Delete Umrah package (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    GroupTicket:
      type: object
      properties:
        id:
          type: integer
        airline:
          type: string
        airlineLogo:
          type: string
        origin:
          type: string
        destination:
          type: string
        country:
          type: string
        travelDate:
          type: string
        seatsAvailable:
          type: integer
        price:
          type: number
        currency:
          type: string
        flightNumber:
          type: string
        category:
          type: string
          enum: [KSA, UAE, Qatar]
        status:
          type: string
          enum: [active, soldout, cancelled]
        createdAt:
          type: string
      required: [id, airline, origin, destination, country, travelDate, seatsAvailable, price, currency, category, status, createdAt]
    GroupTicketInput:
      type: object
      properties:
        airline:
          type: string
        airlineLogo:
          type: string
        origin:
          type: string
        destination:
          type: string
        country:
          type: string
        travelDate:
          type: string
        seatsAvailable:
          type: integer
        price:
          type: number
        currency:
          type: string
        flightNumber:
          type: string
        category:
          type: string
          enum: [KSA, UAE, Qatar]
      required: [airline, origin, destination, country, travelDate, seatsAvailable, price, currency, category]
    GroupTicketUpdate:
      type: object
      properties:
        seatsAvailable:
          type: integer
        price:
          type: number
        status:
          type: string
          enum: [active, soldout, cancelled]
    UmrahTicket:
      type: object
      properties:
        id:
          type: integer
        airline:
          type: string
        airlineLogo:
          type: string
        origin:
          type: string
        travelDate:
          type: string
        returnDate:
          type: string
        seatsAvailable:
          type: integer
        price:
          type: number
        currency:
          type: string
        flightNumber:
          type: string
        status:
          type: string
          enum: [active, soldout, cancelled]
        createdAt:
          type: string
      required: [id, airline, origin, travelDate, returnDate, seatsAvailable, price, currency, status, createdAt]
    UmrahTicketInput:
      type: object
      properties:
        airline:
          type: string
        airlineLogo:
          type: string
        origin:
          type: string
        travelDate:
          type: string
        returnDate:
          type: string
        seatsAvailable:
          type: integer
        price:
          type: number
        currency:
          type: string
        flightNumber:
          type: string
      required: [airline, origin, travelDate, returnDate, seatsAvailable, price, currency]
    UmrahPackage:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        tier:
          type: string
          enum: [economy, vip, luxury]
        hotel:
          type: string
        hotelStars:
          type: integer
        transport:
          type: string
        visaIncluded:
          type: boolean
        duration:
          type: integer
        price:
          type: number
        currency:
          type: string
        description:
          type: string
        amenities:
          type: array
          items:
            type: string
        image:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
      required: [id, name, tier, hotel, hotelStars, transport, visaIncluded, duration, price, currency, description, amenities, status, createdAt]
    UmrahPackageInput:
      type: object
      properties:
        name:
          type: string
        tier:
          type: string
          enum: [economy, vip, luxury]
        hotel:
          type: string
        hotelStars:
          type: integer
        transport:
          type: string
        visaIncluded:
          type: boolean
        duration:
          type: integer
        price:
          type: number
        currency:
          type: string
        description:
          type: string
        amenities:
          type: array
          items:
            type: string
        image:
          type: string
      required: [name, tier, hotel, hotelStars, transport, visaIncluded, duration, price, currency, description, amenities]
...