# lib/api-zod/src/generated/types/umrahPackage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bin Yasin Travels API
 * OpenAPI spec version: 0.1.0
 */
import type { UmrahPackageStatus } from './umrahPackageStatus';
import type { UmrahPackageTier } from './umrahPackageTier';

export interface UmrahPackage {
  id: number;
  name: string;
  tier: UmrahPackageTier;
  hotel: string;
  hotelStars: number;
  transport: string;
  visaIncluded: boolean;
  duration: number;
  price: number;
  currency: string;
  description: string;
  amenities: string[];
  /** @nullable */
  image?: string | null;
  status: UmrahPackageStatus;
  createdAt: string;
}