# lib/api-zod/src/generated/types/umrahTicketInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bin Yasin Travels API
 * OpenAPI spec version: 0.1.0
 */

export interface UmrahTicketInput {
  airline: string;
  airlineLogo?: string;
  origin: string;
  travelDate: string;
  returnDate: string;
  seatsAvailable: number;
  price: number;
  currency: string;
  flightNumber?: string;
}